AppIntro classes: Refactored AppIntro and AppIntro2 to share their common code via AppIntroBase
