+++ library/src/main/java/com/github/paolorotolo/appintro/AppIntro2.java
@@ -1,231 +1,30 @@
 package com.github.paolorotolo.appintro;
 
 import android.animation.ArgbEvaluator;
-import android.os.Build;
 import android.os.Bundle;
-import android.os.Vibrator;
 import android.support.annotation.ColorInt;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
-import android.view.KeyEvent;
 import android.view.View;
-import android.view.Window;
-import android.view.WindowManager;
 import android.widget.FrameLayout;
-import android.widget.ImageView;
-import android.widget.Toast;
 
 import java.util.ArrayList;
-import java.util.List;
-import java.util.Vector;
+public abstract class AppIntro2 extends AppIntroBase {
+    private static final String TAG = "AppIntro2";
 
-public abstract class AppIntro2 extends AppCompatActivity {
-    public final static int DEFAULT_COLOR = 1;
-    private static final int DEFAULT_SCROLL_DURATION_FACTOR = 1;
     private boolean STATUS_BAR_VISIBLE = false;
 
-    protected PagerAdapter mPagerAdapter;
-    protected AppIntroViewPager pager;
-    protected List<Fragment> fragments = new Vector<>();
-    protected List<ImageView> dots;
-    protected int slidesNumber;
-    protected Vibrator mVibrator;
-    protected IndicatorController mController;
-    protected boolean isVibrateOn = false;
-    protected int vibrateIntensity = 20;
-    protected boolean baseProgressButtonEnabled = true;
-    protected boolean progressButtonEnabled = true;
-    protected int selectedIndicatorColor = DEFAULT_COLOR;
-    protected int unselectedIndicatorColor = DEFAULT_COLOR;
-    protected View nextButton;
-    protected View doneButton;
     protected View customBackgroundView;
     protected FrameLayout backgroundFrame;
-    protected int savedCurrentItem;
-    protected ArrayList<PermissionObject> permissionsArray = new ArrayList<>();
-    private static final int PERMISSIONS_REQUEST_ALL_PERMISSIONS = 1;
     private ArrayList<Integer> transitionColors;
     private ArgbEvaluator argbEvaluator = new ArgbEvaluator();
 
-    static enum TransformType {
-        FLOW,
-        DEPTH,
-        ZOOM,
-        SLIDE_OVER,
-        FADE
+    public AppIntro2() {
+        super(R.layout.intro_layout2);
     }
 
     @Override
     final protected void onCreate(Bundle savedInstanceState) {
-        requestWindowFeature(Window.FEATURE_NO_TITLE);
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.intro_layout2);
-
-        nextButton = findViewById(R.id.next);
-        doneButton = findViewById(R.id.done);
         backgroundFrame = (FrameLayout) findViewById(R.id.background);
-        mVibrator = (Vibrator) this.getSystemService(VIBRATOR_SERVICE);
-        mPagerAdapter = new PagerAdapter(getSupportFragmentManager(), fragments);
-        pager = (AppIntroViewPager) findViewById(R.id.view_pager);
-        pager.setAdapter(mPagerAdapter);
-
-        if (savedInstanceState != null) {
-            restoreLockingState(savedInstanceState);
-        }
-
-        nextButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(@NonNull View v) {
-                if (isVibrateOn) {
-                    mVibrator.vibrate(vibrateIntensity);
-                }
-
-                boolean requestPermission = false;
-                int position = 0;
-
-                for (int i = 0; i < permissionsArray.size(); i++) {
-                    requestPermission = pager.getCurrentItem() + 1 == permissionsArray.get(i).getPosition();
-                    position = i;
-                    break;
-                }
-
-                if (requestPermission) {
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                        requestPermissions(permissionsArray.get(position).getPermission(), PERMISSIONS_REQUEST_ALL_PERMISSIONS);
-                        permissionsArray.remove(position);
-                    } else {
-                        pager.setCurrentItem(pager.getCurrentItem() + 1);
-                        onNextPressed();
-                    }
-                } else {
-                    pager.setCurrentItem(pager.getCurrentItem() + 1);
-                    onNextPressed();
-                }
-            }
-        });
-
-        doneButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(@NonNull View v) {
-                if (isVibrateOn) {
-                    mVibrator.vibrate(vibrateIntensity);
-                }
-                onDonePressed();
-            }
-        });
-
-        /**
-         *  ViewPager.setOnPageChangeListener is now deprecated. Use addOnPageChangeListener() instead of it.
-         */
-        pager.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
-            @Override
-            public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
-
-                if (transitionColors != null) {
-                    if(position < (pager.getAdapter().getCount() -1) && position < (transitionColors.size() - 1)) {
-                        pager.setBackgroundColor((Integer) argbEvaluator.evaluate(positionOffset, transitionColors.get(position), transitionColors.get(position + 1)));
-                    } else {
-                        pager.setBackgroundColor(transitionColors.get(transitionColors.size() - 1));
-                    }
-                }
-            }
-
-
-            @Override
-            public void onPageSelected(int position) {
-                if (slidesNumber > 1)
-                    mController.selectPosition(position);
-
-                // Allow the swipe to be re-enabled if a user swipes to a previous slide. Restore
-                // state of progress button depending on global progress button setting
-                if (!pager.isNextPagingEnabled()) {
-                    if (pager.getCurrentItem() != pager.getLockPage()) {
-                        setProgressButtonEnabled(baseProgressButtonEnabled);
-                        pager.setNextPagingEnabled(true);
-                    } else {
-                        setProgressButtonEnabled(progressButtonEnabled);
-                    }
-                } else {
-                    setProgressButtonEnabled(progressButtonEnabled);
-                }
-                onSlideChanged();
-            }
-
-            @Override
-            public void onPageScrollStateChanged(int state) {
-            }
-        });
-
-        setScrollDurationFactor(DEFAULT_SCROLL_DURATION_FACTOR);
-
-        pager.setCurrentItem(savedCurrentItem); //required for triggering onPageSelected for first page
-
-        init(savedInstanceState);
-        slidesNumber = fragments.size();
-
-        if (slidesNumber == 1) {
-            setProgressButtonEnabled(progressButtonEnabled);
-        } else {
-            initController();
-        }
-    }
-
-    protected void setScrollDurationFactor(int factor) {
-        pager.setScrollDurationFactor(factor);
-    }
-
-    @Override
-    protected void onSaveInstanceState(Bundle outState) {
-        super.onSaveInstanceState(outState);
-        outState.putBoolean("baseProgressButtonEnabled", baseProgressButtonEnabled);
-        outState.putBoolean("progressButtonEnabled", progressButtonEnabled);
-        outState.putBoolean("nextEnabled", pager.isPagingEnabled());
-        outState.putBoolean("nextPagingEnabled", pager.isNextPagingEnabled());
-        outState.putInt("lockPage", pager.getLockPage());
-        outState.putInt("currentItem", pager.getCurrentItem());
-    }
-
-
-    protected void restoreLockingState(Bundle savedInstanceState) {
-        super.onRestoreInstanceState(savedInstanceState);
-        this.baseProgressButtonEnabled = savedInstanceState.getBoolean("baseProgressButtonEnabled");
-        this.progressButtonEnabled = savedInstanceState.getBoolean("progressButtonEnabled");
-        this.savedCurrentItem = savedInstanceState.getInt("currentItem");
-        pager.setPagingEnabled(savedInstanceState.getBoolean("nextEnabled"));
-        pager.setNextPagingEnabled(savedInstanceState.getBoolean("nextPagingEnabled"));
-        pager.setLockPage(savedInstanceState.getInt("lockPage"));
-    }
-
-    public AppIntroViewPager getPager() {
-        return pager;
-    }
-
-    private void initController() {
-        if (mController == null)
-            mController = new DefaultIndicatorController();
-
-        FrameLayout indicatorContainer = (FrameLayout) findViewById(R.id.indicator_container);
-        indicatorContainer.addView(mController.newInstance(this));
-
-        mController.initialize(slidesNumber);
-        if (selectedIndicatorColor != DEFAULT_COLOR)
-            mController.setSelectedIndicatorColor(selectedIndicatorColor);
-        if (unselectedIndicatorColor != DEFAULT_COLOR)
-            mController.setUnselectedIndicatorColor(unselectedIndicatorColor);
-    }
-
-    public void addSlide(@NonNull Fragment fragment) {
-        fragments.add(fragment);
-        mPagerAdapter.notifyDataSetChanged();
-    }
-
-    @NonNull
-    public List<Fragment> getSlides() {
-        return mPagerAdapter.getFragments();
     }
 
     /**
@@ -238,189 +37,11 @@ public abstract class AppIntro2 extends AppCompatActivity {
         setProgressButtonEnabled(showDone);
     }
 
-    /**
-     * Setting to to display or hide the Next or Done button. This is a static setting and
-     * button state is maintained across slides until explicitly changed.
-     *
-     * @param progressButtonEnabled Set true to display. False to hide.
-     */
-    public void setProgressButtonEnabled(boolean progressButtonEnabled) {
-        this.progressButtonEnabled = progressButtonEnabled;
-        if (progressButtonEnabled) {
-            if (pager.getCurrentItem() == slidesNumber - 1) {
-                setButtonState(nextButton, false);
-                setButtonState(doneButton, true);
-            } else {
-                setButtonState(nextButton, true);
-                setButtonState(doneButton, false);
-            }
-        } else {
-            setButtonState(nextButton, false);
-            setButtonState(doneButton, false);
-        }
-    }
-
-    public boolean isProgressButtonEnabled() {
-        return progressButtonEnabled;
-    }
-
-    private void setButtonState(View button, boolean show) {
-        if (show) {
-            button.setVisibility(View.VISIBLE);
-        } else {
-            button.setVisibility(View.INVISIBLE);
-        }
-    }
-
-    public void setNavBarColor(String Color) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            getWindow().setNavigationBarColor(android.graphics.Color.parseColor(Color));
-        }
-    }
-
-    public void showStatusBar(boolean isVisible) {
-        this.STATUS_BAR_VISIBLE = isVisible;
-
-        if (STATUS_BAR_VISIBLE) {
-            getWindow().clearFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN);
-        } else {
-            getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
-                    WindowManager.LayoutParams.FLAG_FULLSCREEN);
-        }
-    }
     public void setBackgroundView(View view){
         customBackgroundView = view;
         if (customBackgroundView!=null){
             backgroundFrame.addView(customBackgroundView);
         }
-    }
-
-    public void setVibrate(boolean vibrate) {
-        this.isVibrateOn = vibrate;
-    }
-
-    public void setVibrateIntensity(int intensity) {
-        this.vibrateIntensity = intensity;
-    }
-
-    public void setFadeAnimation() {
-        pager.setPageTransformer(true, new ViewPageTransformer(ViewPageTransformer.TransformType.FADE));
-    }
-
-    public void setZoomAnimation() {
-        pager.setPageTransformer(true, new ViewPageTransformer(ViewPageTransformer.TransformType.ZOOM));
-    }
-
-    public void setFlowAnimation() {
-        pager.setPageTransformer(true, new ViewPageTransformer(ViewPageTransformer.TransformType.FLOW));
-    }
-
-    public void setSlideOverAnimation() {
-        pager.setPageTransformer(true, new ViewPageTransformer(ViewPageTransformer.TransformType.SLIDE_OVER));
-    }
-
-    public void setDepthAnimation() {
-        pager.setPageTransformer(true, new ViewPageTransformer(ViewPageTransformer.TransformType.DEPTH));
-    }
-
-    public void setCustomTransformer(@Nullable ViewPager.PageTransformer transformer) {
-        pager.setPageTransformer(true, transformer);
-    }
-
-    public void setOffScreenPageLimit(int limit) {
-        pager.setOffscreenPageLimit(limit);
-    }
-
-    /**
-     * Set a progress indicator instead of dots. This is recommended for a large amount of slides. In this case there
-     * could not be enough space to display all dots on smaller device screens.
-     */
-    public void setProgressIndicator() {
-        mController = new ProgressIndicatorController();
-    }
-
-    /**
-     * Set a custom {@link IndicatorController} to use a custom indicator view for the {@link AppIntro2} instead of the
-     * default one.
-     *
-     * @param controller The controller to use
-     */
-    public void setCustomIndicator(@NonNull IndicatorController controller) {
-        mController = controller;
-    }
-
-    public abstract void init(@Nullable Bundle savedInstanceState);
-
-    public abstract void onDonePressed();
-
-    public abstract void onNextPressed();
-
-    public abstract void onSlideChanged();
-
-    @Override
-    public boolean onKeyDown(int code, KeyEvent kevent) {
-        if (code == KeyEvent.KEYCODE_ENTER || code == KeyEvent.KEYCODE_BUTTON_A || code == KeyEvent.KEYCODE_DPAD_CENTER) {
-            ViewPager vp = (ViewPager) this.findViewById(R.id.view_pager);
-            if (vp.getCurrentItem() == vp.getAdapter().getCount() - 1) {
-                onDonePressed();
-            } else {
-                vp.setCurrentItem(vp.getCurrentItem() + 1);
-            }
-            return false;
-        }
-        return super.onKeyDown(code, kevent);
-    }
-
-    /**
-     * Set DEFAULT_COLOR for color value if you don't want to change it
-     */
-    public void setIndicatorColor(int selectedIndicatorColor, int unselectedIndicatorColor) {
-        this.selectedIndicatorColor = selectedIndicatorColor;
-        this.unselectedIndicatorColor = unselectedIndicatorColor;
-
-        if (mController != null) {
-            if (selectedIndicatorColor != DEFAULT_COLOR)
-                mController.setSelectedIndicatorColor(selectedIndicatorColor);
-            if (unselectedIndicatorColor != DEFAULT_COLOR)
-                mController.setUnselectedIndicatorColor(unselectedIndicatorColor);
-        }
-    }
-
-    /**
-     * Setting to disable forward swiping right on current page and allow swiping left. If a swipe
-     * left occurs, the lock state is reset and swiping is re-enabled. (one shot disable) This also
-     * hides/shows the Next and Done buttons accordingly.
-     *
-     * @param lockEnable Set true to disable forward swiping. False to enable.
-     */
-    public void setNextPageSwipeLock(boolean lockEnable) {
-        if (lockEnable) {
-            // if locking, save current progress button visibility
-            baseProgressButtonEnabled = progressButtonEnabled;
-            setProgressButtonEnabled(!lockEnable);
-        } else {
-            // if unlocking, restore original button visibility
-            setProgressButtonEnabled(baseProgressButtonEnabled);
-        }
-        pager.setNextPagingEnabled(!lockEnable);
-    }
-
-    /**
-     * Setting to disable swiping left and right on current page. This also
-     * hides/shows the Next and Done buttons accordingly.
-     *
-     * @param lockEnable Set true to disable forward swiping. False to enable.
-     */
-    public void setSwipeLock(boolean lockEnable) {
-        if (lockEnable) {
-            // if locking, save current progress button visibility
-            baseProgressButtonEnabled = progressButtonEnabled;
-            //setProgressButtonEnabled(!lockEnable);
-        } else {
-            // if unlocking, restore original button visibility
-            setProgressButtonEnabled(baseProgressButtonEnabled);
-        }
-        pager.setPagingEnabled(!lockEnable);
     }
 
     /**
@@ -430,32 +51,5 @@ public abstract class AppIntro2 extends AppCompatActivity {
      * */
     public void setAnimationColors(@ColorInt ArrayList<Integer> colors) {
         transitionColors = colors;
-    }
-
-    private static String TAG = "AppIntro2";
-
-    public void askForPermissions(String[] permissions, int slidesNumber) {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-            if (slidesNumber == 0) {
-                Toast.makeText(getBaseContext(), "Invalid Slide Number", Toast.LENGTH_SHORT).show();
-            } else {
-                PermissionObject permission = new PermissionObject(permissions, slidesNumber);
-                permissionsArray.add(permission);
-                setSwipeLock(true);
-            }
-        }
-    }
-
-    public void onRequestPermissionsResult(int requestCode, @NonNull String permissions[],
-                                           @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        switch (requestCode) {
-            case PERMISSIONS_REQUEST_ALL_PERMISSIONS:
-                pager.setCurrentItem(pager.getCurrentItem() + 1);
-                break;
-            default:
-                Log.e(TAG, "Unexpected request code");
-        }
-
     }
 }
